Split tmux pane input queue
